[server] fix LookUp for darwin
ref commint: 4b355ef12576454af7ad1469e3f89368d9461fb1
