+++ server/libs/grpc/utils_linux.go
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2023 Yunshan Networks
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package grpc
+
+import (
+	"fmt"
+	"net"
+
+	"github.com/vishvananda/netlink"
+)
+
+func Lookup(host net.IP) (net.IP, error) {
+	routes, err := netlink.RouteGet(host)
+	if err != nil {
+		return nil, fmt.Errorf("RouteGet %v %s", host, err)
+	}
+	route := routes[0]
+	src := route.Src
+	if route.Src.To4() != nil {
+		src = route.Src.To4()
+	}
+	return src, nil
+}
